+++ Tests/KeyboardKitTests/Keyboard/KeyboardInputViewControllerTests.swift
@@ -122,8 +122,9 @@ class KeyboardInputViewControllerTests: XCTestCase {
     // MARK: - Services
 
     func servicesHaveStandardInstancesByDefault() {
-        XCTAssertNotNil(vc.autocompleteProvider as? DeprecatedAutocompleteProvider)
+        XCTAssertNotNil(vc.autocompleteProvider as? DisabledAutocompleteProvider)
         XCTAssertNotNil(vc.calloutActionProvider as? StandardCalloutActionProvider)
+        XCTAssertNotNil(vc.dictationService as? DisabledKeyboardDictationService)
         XCTAssertNotNil(vc.keyboardActionHandler as? StandardKeyboardActionHandler)
         XCTAssertNotNil(vc.keyboardBehavior as? StandardKeyboardBehavior)
         XCTAssertNotNil(vc.keyboardLayoutProvider as? StandardKeyboardLayoutProvider)